Move the main loop into an extra class
